feat: add example using Sentry V2 SDK
Notes: The V1 example doesn't seem to work for me after recently trying
to set up Sentry. The issue seems to be related to the `warnings` and
`threading` modules that Sentry uses not being included properly into
the Temporal Workflow's sandbox. This results in the Workflow execution
failing to start in the worker that handles the Workflow. I don't see any
issues with the Sentry interceptor on other workers that only handle
activities, so it is definitely related to the Workflow sandbox.

Note: I also tried disabling the worker's sandbox and this seemed to work,
even though you get a sporadic error in the Temporal UI.

The solution seems to be to migrate the interceptor to use V2 of Sentry's
SDK, where the Hub object is now deprecated in favour of using scopes.

It seems that using the scope context manager doesn't have the same issues
with the warnings and threading libs (which it does still use).

I've created a new example to keep the V1 example in place, since V2 only
works with Python 3.6 or above.